Move project files into img-txt_viewer directory
