+++ img-txt_viewer/build_app.bat
