+++ img-txt_viewer/main/dict/dictionary.csv
