+++ img-txt_viewer/main/scripts/OnnxTagger.py
