+++ img-txt_viewer/main/scripts/TagEditor.py
