+++ img-txt_viewer/main/scripts/ThumbnailPanel.py
