+++ img-txt_viewer/main/scripts/about_img_txt_viewer.py
