+++ img-txt_viewer/main/scripts/batch_crop_images.py
