+++ img-txt_viewer/main/scripts/batch_rename.py
