+++ img-txt_viewer/main/scripts/batch_resize_images.py
