+++ img-txt_viewer/main/scripts/calculate_file_stats.py
