+++ img-txt_viewer/main/scripts/edit_panel.py
