+++ img-txt_viewer/main/scripts/entry_helper.py
