+++ img-txt_viewer/main/scripts/filter_csv.py
