+++ img-txt_viewer/main/scripts/find_replace_widget.py
