+++ img-txt_viewer/main/scripts/resize_image.py
