+++ img-txt_viewer/main/scripts/settings_manager.py
