+++ img-txt_viewer/main/scripts/text_controller.py
