+++ img-txt_viewer/main/scripts/video_player_widget.py
