+++ img-txt_viewer/main/scripts/video_thumbnail_generator.py
